TEMPORARY navigation complete
This is NOT what the final navigation will look like. I'm just using the drawer on the side as means of debugging each screen so we can easily see what the heck they look like. THIS IT NOT FINAL and yet I know someone is going to get all bent out of shape over it.